refactor: rename hyperv-winrm package to hyperv
