+++ api/hyperv/vm_switch.go
@@ -1,4 +1,4 @@
-package hyperv_winrm
+package hyperv
 
 import (
 	"context"
